backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
openai==1.56.0  
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
python-docx==1.1.2
python-pptx==0.6.23 
python-multipart==0.0.9
httpx==0.28.0  